* renaming vz::primitive to vz::geometrics (file name is also changed to "Geometrics.h) * remove 'bindless' material index from ShaderGeometryPart * add 'bindless' material indices to ShaderMeshInstance (constraints: MAXPARTS 16)

+++ EngineCore/Common/RenderPath3D.cpp
@@ -48,8 +48,8 @@ namespace vz
 		{
 			viewport.top_left_x = 0;
 			viewport.top_left_y = 0;
-			viewport.width = width_;
-			viewport.height = height_;
+			viewport.width = (float)width_;
+			viewport.height = (float)height_;
 		}
 		//scissor
 		if (!useManualSetScissor)
